Fix ASN data type
The BGP AS Number size was specified as an Integer, which
supports signed numbers from -2,147,483,648 to 2,147,483,647.
However, BGP ASNs use unsigned 32-bit numbers, which range
from 0 to 4,294,967,295. This patch changes the column from
an Integer to a BigInteger, so that the full ASN range can
be supported.

Fixes #886

+++ aim/db/models.py
@@ -1100,7 +1100,7 @@ class L3OutInterfaceBgpPeerP(model_base.Base, model_base.HasAimId,
     interface_profile_name = model_base.name_column(nullable=False)
     interface_path = sa.Column(VARCHAR(512, charset='latin1'), nullable=False)
     addr = sa.Column(sa.String(64), nullable=False)
-    asn = sa.Column(sa.Integer)
+    asn = sa.Column(sa.BigInteger)
 
     def to_attr(self, session):
         res_attr = super(L3OutInterfaceBgpPeerP, self).to_attr(session)
